Move dates utilities to common directory
